Implement very crude menu manager and demo.
This is a little visually buggy (the cursor can XOR itself away when you hit the bounds of the active menu).
It's stupid and messy, but internally it works perfect. There are dozens of comments about ideas. I'll
get rid of them in the next few commits, but I'm keeping them here for now...