Fix keyboard bugs, make demo prettier, small housekeeping
Fixed the keyboard reliability bug where key presses from other buttons would trigger _keytest.
The original vectors for interrupts 1 and 5 interfere with keyboard. So we map interrupt 1 to greyscale routines _only_, and disable interrupt 5 entirely.

Demo is prettier now, with 8px cursor option, help text, and bigger text.

Rewrote the keyboard routine to use _keytest_optimized. Not too certain of the codegen, I'll analyze it later.
I think it may be best for me to just write the routine myself with _rowread though with a switch(case) on the column of the 2nd/Arrows row.

Removed the OS timer debounce routine, it requires AutoInt5, which we can't reliably use. So I switched to a nop loop for now.

clangd gets upset about an lvalue cast (among other things). This is technically not allowed in (modern?) C, but TIGCC (GCC4.1) patched it in, so it's okay.

Lastly, small housekeeping: clangd stops showing inlay designators because I don't like them.