Another MenuManager overhaul, also add textbox functionality
The MenuManager no longer fires callbacks on its own. It will return a menu item,
and you can choose to fire the callback yourself. A helpful macro that mimics
the old functionality of MenuManager is provided in macros.h

We make this change to eliminate callstack badness. Before, if you wanted to launch a game
from the main menu, you would have to set that as a callback. I obviously don't want
the main game loop to be called from atop the StartMenuManager stack frame! But I also
don't quite want to implement SL/LJ or inline assembler to modify the stack frame
and start executing the callback function in place. This method of returning
does have the detriment of needing a branch or two (check if we returned something),
and is not quite as "interrupt-like" as I'd hoped, but it is better overall, and keeps
our stack frames flat.

This destroys what little nested menuing functionality we had, but we wil leave that up to the programmer.
After all, even with the old MenuManager, nested menu functionality was strange, and we had no real way
to perform redraw automatically. In this way, we can leave the whole redraw loop/nesting loop up to the programmer.
MenuManager is there ONLY to handle keyboard input for one menu layer, and can be reset for different menu layers
if needed.

Textbox functionality was added. It uses interrupt 5 and OS routines, so its a little buggy with rowread.
It'll work for now. Wraps text correctly as far as I can tell, but only prints one line for now.

There's also a bit of a bug/oversight with 8px borders though, where they can't quite touch the edge of the screen. TODO?